Разделил все на 3 файла: model, view, controller (по паттерну MVC)
