+++ app/scripts/model.js
@@ -49,16 +49,6 @@ export class User {
     }
 }
 
-// -----------------------------------------------------------------------------------
-// Функция проверяет - встречается ли подстрока chunk в строке full без учета регистра
-// -----------------------------------------------------------------------------------
-export function isMatching(full, chunk) {
-    let fullStr = full.toLowerCase(),
-        chunkStr = chunk.toLowerCase();
-
-    return fullStr.indexOf(chunkStr) !== -1;
-}
-
 // -----------------------------------------------
 // Функция отправки текстового сообщения на сервер
 // -----------------------------------------------
